Do coding for lesson 5a
Since I want to run both the frontend and the backend locally, and be 
able to change the frontend on-the-fly, I have to remove express.static 
from the backend (NOT providing the frontend from static content), and 
add the proxy "setting" to the package.json of the frontend, so that 
when it is running on port 3000, it's backend calls (to, e.g. 
3000/api/notes), will instead be routed to 3001/api/notes, where the 
backend is running.