Update `user` and `blogs` fields at init
Also add a test for `user` field existence in blogs and fix an issue
that arose after adding the user info to init. Previously there was no
user info in our initial blogs, so fetching one and editing it didn't
reveal the problem that arose when the user field got populated. During
a `put` operation we have to "depopulate" the populated field.