Do task 7.19 (7.18 was done in previous commit)
NOT completely satisfied with this solution. Wondering about:
- a) having such a hard time with forms: my first attempt had some
issue with conflicting "controlled" vs. "uncontrolled", which I did not
understand and didn't bother to figure out. My current solution has,
I feel, way too much complexity for what it does.
- b) would it make more sense for the addComment API to return the
commented blog, rather than the comment. AND if I did it that way,
should I return a 200 instead of 201? (200 being what we currently
return for a blog update. That would certainly make use on the
frontend easier, and would reduce the number of API calls. That would totally
seem to make sense, but currently just trying to slog my way
thru to the end of this chapter, so very much in a "I don't care"
state-of-mind.